Backend/requirements.txt
Django==3.1.2
djangorestframework==3.12.1
psycopg2-binary==2.8.6
dj-database-url==0.5.0
gunicorn==20.0.4
firebase-admin==4.4.0